requirements.txt
mysqlclient==2.1.1
pymysql==1.0.2
python-dotenv==0.19.2
regex==2021.7.6
requests==2.28.1
fastapi==0.88.0
uvicorn==0.20.0
python-multipart==0.0.5